% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{carros_eletronicas}
\alias{carros_eletronicas}
\title{Infrações cometidas por carros capturadas eletronicamente (radar)}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{data}{data da infração}
  \item{hora}{hora do dia da infração}
  \item{enquadramento}{motivo da multa}
  \item{local}{local da infração}
  \item{qtd}{quantidade de infrações}
}}
\source{
\url{http://mobilidadesegura.prefeitura.sp.gov.br}
}
\usage{
carros_eletronicas
}
\description{
Banco de dados extraído manualmente do portal da mobilidade urbana.
\url{http://mobilidadesegura.prefeitura.sp.gov.br}
}
\keyword{datasets}
